Extract lib
